requirements.txt
# Core dependencies
streamlit>=1.28.0
sentence-transformers>=2.2.2
faiss-cpu>=1.7.4
rank-bm25>=0.2.2
transformers>=4.35.0
torch>=2.0.0

# Wikipedia data collection
wikipedia-api>=0.6.0
beautifulsoup4>=4.12.0
requests>=2.31.0

# NLP & Text Processing
nltk>=3.8.1
tiktoken>=0.5.0

# Evaluation metrics
rouge-score>=0.1.2
bert-score>=0.3.13
scikit-learn>=1.3.0
numpy>=1.24.0
pandas>=2.0.0

# Visualization & Reporting
plotly>=5.18.0
matplotlib>=3.7.0
seaborn>=0.12.0
fpdf2>=2.7.0

# Utilities
tqdm>=4.66.0
python-dotenv>=1.0.0
